sked snapshot 2022-01-06 (IVS sked_cat skedall_2022-01-06.tgz)
Reconstructed from the IVS Analysis Coordinator sked catalog snapshot skedall_2022-01-06.tgz. Source, config and scripts only; build artifacts and binaries stripped. Date is the snapshot's nominal (filename) date, not per-change.

+++ sked/display_scan_info.f
@@ -12,7 +12,7 @@
 
 ! functions
       real*8 speed 
-      integer isecdif
+      integer*4 isecdif
 
 ! Input variables
       integer mjd                    !time of scan.
